// src/app/products/[id]/page.tsx
import ProductInteraction from "@/components/Productinteraction";
import { ProductsType } from "@/types";
import Image from "next/image";

// TEMPORARY
const product: ProductsType = [
	{
		id: 1,
		name: "Adidas CoreFit T-Shirt",
		shortDescription:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		description:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		price: 39.9,
		sizes: ["s", "m", "l", "xl", "xxl"],
		colors: ["gray", "purple", "green"],
		images: {
			gray: "/products/1g.png",
			purple: "/products/1p.png",
			green: "/products/1gr.png",
		},
	},
	{
		id: 2,
		name: "Puma Ultra Warm Zip",
		shortDescription:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		description:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		price: 59.9,
		sizes: ["s", "m", "l", "xl"],
		colors: ["gray", "green"],
		images: { gray: "/products/2g.png", green: "/products/2gr.png" },
	},
	{
		id: 3,
		name: "Nike Air Essentials Pullover",
		shortDescription:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		description:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		price: 69.9,
		sizes: ["s", "m", "l"],
		colors: ["green", "blue", "black"],
		images: {
			green: "/products/3gr.png",
			blue: "/products/3b.png",
			black: "/products/3bl.png",
		},
	},
	{
		id: 4,
		name: "Nike Dri Flex T-Shirt",
		shortDescription:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		description:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		price: 29.9,
		sizes: ["s", "m", "l"],
		colors: ["white", "pink"],
		images: { white: "/products/4w.png", pink: "/products/4p.png" },
	},
	{
		id: 5,
		name: "Under Armour StormFleece",
		shortDescription:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		description:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		price: 49.9,
		sizes: ["s", "m", "l"],
		colors: ["red", "orange", "black"],
		images: {
			red: "/products/5r.png",
			orange: "/products/5o.png",
			black: "/products/5bl.png",
		},
	},
	{
		id: 6,
		name: "Nike Air Max 270",
		shortDescription:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		description:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		price: 59.9,
		sizes: ["40", "42", "43", "44"],
		colors: ["gray", "white"],
		images: { gray: "/products/6g.png", white: "/products/6w.png" },
	},
	{
		id: 7,
		name: "Nike Ultraboost Pulse ",
		shortDescription:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		description:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		price: 69.9,
		sizes: ["40", "42", "43"],
		colors: ["gray", "pink"],
		images: { gray: "/products/7g.png", pink: "/products/7p.png" },
	},
	{
		id: 8,
		name: "Levi’s Classic Denim",
		shortDescription:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		description:
			"Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit. Lorem ipsum dolor sit amet consect adipisicing elit lorem ipsum dolor sit.",
		price: 59.9,
		sizes: ["s", "m", "l"],
		colors: ["blue", "green"],
		images: { blue: "/products/8b.png", green: "/products/8gr.png" },
	},
];

export const generateMetadata = async ({
	params,
}: {
	params: { id: string };
}) => {
	// TODO:get the product from db
	// TEMPORARY
	const selectedProduct = product.find((p) => p.id === parseInt(params.id));
	return {
		title: selectedProduct?.name || "Product",
		describe: selectedProduct?.description || "",
	};
};

const ProductPage = async ({
	params,
	searchParams,
}: {
	params: Promise<{ id: string }>;
	searchParams: Promise<{ color: string; size: string }>;
}) => {
	const { size, color } = await searchParams;

	const selectedProduct = product[0];
	const selectedSize = size || (selectedProduct.sizes[0] as string);
	const selectedColor = color || (selectedProduct.colors[0] as string);
	return (
		<div className="flex flex-col gap-4 lg:flex-row md:gap-12 mt-12">
			{/* IMAGE */}
			<div className="w-full lg:w-5/12 relative aspect-[2/3]">
				<Image
					src={selectedProduct.images[selectedColor]}
					alt={selectedProduct.name}
					fill
					className="object-contain rounded-md"
				/>
			</div>
			{/* DETAILS */}
			<div className="w-full lg:w-7/12 flex flex-col gap-4">
				<h1 className="text-2xl font-medium">{selectedProduct.name}</h1>
				<p className="text-gray-500">{selectedProduct.description}</p>
				<h2 className="text-2xl font-semibold">${selectedProduct.price.toFixed(2)}</h2>
				<ProductInteraction
					product={selectedProduct}
					selectedSize={selectedSize}
					selectedColor={selectedColor}
				/>
				{/* CARD INFO */}
				<div className="flex items-center gap-2 mt-4">
					<Image
						src="/klarna.png"
						alt="klarna"
						width={50}
						height={25}
						className="rounded-md"
					/>
					<Image
						src="/cards.png"
						alt="cards"
						width={50}
						height={25}
						className="rounded-md"
					/>
					<Image
						src="/stripe.png"
						alt="stripe"
						width={50}
						height={25}
						className="rounded-md"
					/>
				</div>
				<p className="text-gray-500 text-xs">
					By clicking Pay Now, you agree to our{" "}
					<span className="underline hover:text-black">Terms & Conditions</span>{" "}
					and <span className="underline hover:text-black">Privacy Policy</span>
					. You authorize us to charge your selected payment method for the
					total amount shown. All sales are subject to our return and{" "}
					<span className="underline hover:text-black">Refund Policies</span>.
				</p>
			</div>
		</div>
	);
};

export default ProductPage;
